Structural and pause Restart
I have moved around a couple things nothing huge. it still works as well as before.
I have also included a reset for pause however having problems with map to perfectly reset :/
we need to also make sure that when the explosion occurs we get rid of the runner cos she can still be seen falling.